[Chore] 폴더링 및 환경 설정
