--- EatsOrder/ContentView.swift
@@ -1,24 +1,0 @@
-//
-//  ContentView.swift
-//  EatsOrder
-//
-//  Created by 한수빈 on 5/9/25.
-//
-
-import SwiftUI
-
-struct ContentView: View {
-    var body: some View {
-        VStack {
-            Image(systemName: "globe")
-                .imageScale(.large)
-                .foregroundStyle(.tint)
-            Text("Hello, world!")
-        }
-        .padding()
-    }
-}
-
-#Preview {
-    ContentView()
-}
