+++ EatsOrder/Resource/Supporting Files/Info.plist
@@ -1,5 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
-<dict/>
+<dict>
+	<key>SESAC_KEY</key>
+	<string>$(SESAC_KEY)</string>
+</dict>
 </plist>
