+++ EatsOrder/Source/ContentView.swift
@@ -1,0 +1,24 @@
+//
+//  ContentView.swift
+//  EatsOrder
+//
+//  Created by 한수빈 on 5/9/25.
+//
+
+import SwiftUI
+
+struct ContentView: View {
+  var body: some View {
+    VStack {
+      Image(systemName: "globe")
+        .imageScale(.large)
+        .foregroundStyle(.tint)
+      Text(Environments.apiKey)
+    }
+    .padding()
+  }
+}
+
+#Preview {
+  ContentView()
+}
