+++ EatsOrder/Source/EatsOrderApp.swift
@@ -9,9 +9,9 @@ import SwiftUI
 
 @main
 struct EatsOrderApp: App {
-    var body: some Scene {
-        WindowGroup {
-            ContentView()
-        }
+  var body: some Scene {
+    WindowGroup {
+      ContentView()
     }
+  }
 }
