+++ EatsOrder/Source/Environments.swift
@@ -1,0 +1,14 @@
+//
+//  Environments.swift
+//  EatsOrder
+//
+//  Created by 한수빈 on 5/9/25.
+//
+
+import Foundation
+
+enum Environments {
+  static var apiKey: String {
+    Bundle.main.object(forInfoDictionaryKey: "SESAC_KEY") as! String
+  }
+}
